+++ lib/core/service_locator.dart
@@ -21,6 +21,7 @@ import 'package:quizdy/data/repositories/ai/ai_repository_factory.dart';
 import 'package:quizdy/data/services/ai/ai_document_chunking_service.dart';
 import 'package:quizdy/data/services/ai/ai_jit_processing_service.dart';
 import 'package:quizdy/data/services/ai/ai_question_generation_service.dart';
+import 'package:quizdy/data/services/app_remote_config_service.dart';
 import 'package:quizdy/data/services/configuration_service.dart';
 import 'package:quizdy/domain/use_cases/check_file_changes_use_case.dart';
 
@@ -60,6 +61,10 @@ class ServiceLocator {
 
     final configurationService = getIt.registerSingleton<ConfigurationService>(
       ConfigurationService(sharedPreferences: sharedPreferences),
+    );
+
+    getIt.registerSingleton<AppRemoteConfigService>(
+      AppRemoteConfigService(sharedPreferences: sharedPreferences),
     );
 
     getIt.registerSingleton<AiRepositoryFactory>(